First parser prototype
Self-contained parser using parser combinators.

Checking in my progress so I experiment more easily without losing work.
The parser kinda works but there are bugs. One of the bugs is badly
formatted predicate arguments will fail on end of file instead of on
argument parsing.

The rule production was tricky. The one combinator is eager, so the
first parse that succeeds is kept. Since the atom production is a prefix
of the rule production, if the atom production is first in alternation
it'll match when a rule should have matched. Checking for a rule before
a fact fixes it.