Add first version of README
The README is as much of a design wishlist as it is documentation. The
README was being written as the code was, usually before the code was,
in the style of readme-driven development.

Some things aren't done yet, like the repl, and the cabin command is
actually `python cli.py`. Some module member names are also different.