Only bind body variables in Tp operator
The immediate consequence operator only needs to bind variables from the
body because a rule must be range restricted to be valid, i.e., all the
variables in the head must also be in the body. So the function only
needs to use the variables from the body, because the variables from the
head are guaranteed to be there.