Add interactive mode to cli
Use the -i switch to enter the interactive interpreter.

If a file is provided along with the -i switch, the file will be
evaluated first and then the repl will start. The program from the file
and the rules entered into the repl will share the same engine instance,
so results from the program in the file can be interactively queried.

Makes it so that file no longer defaults to stdin.